+++ rector.php
@@ -4,27 +4,19 @@ declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
 use Rector\Php73\Rector\String_\SensitiveHereNowDocRector;
-use Rector\Set\ValueObject\DowngradeLevelSetList;
-use Rector\Set\ValueObject\LevelSetList;
-use Rector\Set\ValueObject\SetList;
 
-return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->paths([
+return RectorConfig::configure()
+    ->withPaths([
         __DIR__ . '/src',
-    ]);
-
-    $rectorConfig->parallel(300);
-
-    $rectorConfig->skip([
+    ])
+    ->withParallel(timeoutSeconds: 300)
+    ->withSkip([
         SensitiveHereNowDocRector::class,
-    ]);
-
-    // define sets of rules
-    $rectorConfig->sets([
-        SetList::NAMING,
-        SetList::CODE_QUALITY,
-        SetList::CODING_STYLE,
-        LevelSetList::UP_TO_PHP_81,
-        DowngradeLevelSetList::DOWN_TO_PHP_81,
-    ]);
-};
+    ])
+    ->withPreparedSets(
+        naming: true,
+        codeQuality: true,
+        codingStyle: true,
+    )
+    ->withPhpSets(php82: true)
+    ->withDowngradeSets(php82: true);